separate bvhcast for future investigation
