+++ src/core/MeshBVH.js
@@ -453,114 +453,6 @@ export class MeshBVH {
 
 	}
 
-	bvhcast( otherBvh, matrixToLocal, callbacks ) {
-
-		let {
-			intersectsRange,
-			intersectsTriangle,
-		} = callbacks;
-
-		const geometry = otherBvh.geometry;
-		const indexAttr = geometry.index;
-		const positionAttr = geometry.attributes.position;
-
-		const triangle = trianglePool.getPrimitive();
-		const triangle2 = trianglePool.getPrimitive();
-		if ( intersectsTriangle ) {
-
-			function iterateOverDoubleTriangles( offset1, count1, offset2, count2, depth1, index1, depth2, index2 ) {
-
-				for ( let i2 = offset2, l2 = offset2 + count2; i2 < l2; i2 ++ ) {
-
-					setTriangle( triangle2, i2 * 3, indexAttr, positionAttr );
-					triangle2.a.applyMatrix4( matrixToLocal );
-					triangle2.b.applyMatrix4( matrixToLocal );
-					triangle2.c.applyMatrix4( matrixToLocal );
-					triangle2.needsUpdate = true;
-
-					for ( let i1 = offset1, l1 = offset1 + count1; i1 < l1; i1 ++ ) {
-
-						setTriangle( triangle, i1 * 3, indexAttr, positionAttr );
-						triangle.needsUpdate = true;
-
-						if ( intersectsTriangle( triangle, triangle2, i1, i2, depth1, index1, depth2, index2 ) ) {
-
-							return true;
-
-						}
-
-					}
-
-				}
-
-				return false;
-
-			}
-
-			if ( intersectsRange ) {
-
-				const originalIntersectsRange = intersectsRange;
-				intersectsRange = function ( offset1, count1, offset2, count2, depth1, index1, depth2, index2 ) {
-
-					if ( ! originalIntersectsRange( offset1, count1, offset2, count2, depth1, index1, depth2, index2 ) ) {
-
-						return iterateOverDoubleTriangles( offset1, count1, offset2, count2, depth1, index1, depth2, index2 );
-
-					}
-
-					return true;
-
-				};
-
-			} else {
-
-				intersectsRange = iterateOverDoubleTriangles;
-
-			}
-
-		}
-
-		let result = false;
-		let byteOffset = 0;
-		let byteOffset2 = 0;
-		for ( const root of this._roots ) {
-
-			setBuffer( root );
-
-			for ( const root2 of otherBvh._roots ) {
-
-				setBuffer2( root2 );
-				result = bvhcast( 0, 0, matrixToLocal, intersectsRange, byteOffset, byteOffset2 );
-				clearBuffer2();
-
-				if ( result ) {
-
-					break;
-
-				}
-
-				byteOffset2 += root2.byteLength;
-
-			}
-
-			clearBuffer();
-
-			if ( result ) {
-
-				break;
-
-			}
-
-			byteOffset += root.byteLength;
-
-		}
-
-		trianglePool.releasePrimitive( triangle );
-		trianglePool.releasePrimitive( triangle2 );
-		return result;
-
-	}
-
 	shapecast( callbacks, _intersectsTriangleFunc, _orderNodesFunc ) {
 
 		const geometry = this.geometry;
